pyproject.toml
[tool.poetry]
name = "archie-cli"
version = "0.1.0"
description = "TP-Link Archer C7 CLI"
authors = ["Omer Aloni <omeraloni@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.24.0"
click = "^7.1.2"
coloredlogs = "^14.0"
python-crontab = "^2.5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
